+++ CHANGELOG.md
@@ -19,7 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BSplineCurve3D: trim
 - FullArc3D: hash
 - SphericalSurface3D: enhance repair_periodicity_method
-
+- CylindricalSurface3D: concurrent_plane_intersection
+- 
 ### Refactor
 - refator some classes' init in primitives3D. 
 - Shells: refactor.